Added more missing item types
